% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calmean.R
\name{cal_mean}
\alias{cal_mean}
\title{Calculate the mean value for different groups}
\usage{
cal_mean(data, group, target)
}
\arguments{
\item{data}{is the data table we will focus on}

\item{group}{is the property that user would like to group by.}

\item{target}{is the target column that we want to calculate its mean value}
}
\value{
the mean values of different groups
}
\description{
Calculate the mean value for different groups
}
\details{
the function is aimed to firstly group the data by a property and then calculate the mean values of the grouped data.
}
